refactor: 재사용 요소 분리 (#262)
