+++ src/features/_wide/layout/footer/1-explore-contents/ExploreContents.tsx
@@ -1,0 +1,15 @@
+import FooterLink from "../footer-link/FooterLink"
+
+const ExploreContents = () => {
+  return (
+    <>
+      <h3 className="text-text-disabled text-sm">EXPLORE</h3>
+      <FooterLink to="/">About</FooterLink>
+      <FooterLink to="/">Scent Recommendation</FooterLink>
+      <FooterLink to="/">Feed</FooterLink>
+      <FooterLink to="/">My Page</FooterLink>
+    </>
+  )
+}
+
+export default ExploreContents
