+++ src/features/_wide/layout/footer/Footer.tsx
@@ -1,6 +1,6 @@
 import { useMediaQuery } from "@/shared/utils/use-media-query"
-import NarrowFooter from "./NarrowFooter"
-import WideFooter from "./WideFooter"
+import NarrowFooter from "./narrow-footer/NarrowFooter"
+import WideFooter from "./wide-footer/WideFooter"
 
 const Footer = () => {
   const { isWide } = useMediaQuery()
